Update draft-pignataro-enviro-sustainability-architecture.xml
Various text edits and additional text snippets.  In the introduction setting, added explanation of why sustainability may involve tradeoffs to provide a stronger motivation.  In section 2.2, added greater explanation of the tradeoffs between sustainability and other goals, which might result in different operational choices, and the role of the operator to navigate those tradeoffs.  Folded section 2.7.1 (an "orphaned" subsection) into 2.7 and expanded on the role of network management to provide visibility.